1.3.8

+++ handlers/my-stats.js
@@ -5,17 +5,13 @@ const { userName } = require('../utils')
 
 module.exports = async (ctx) => {
   if (['supergroup', 'group'].includes(ctx.chat.type)) {
-    const message = await ctx.replyWithHTML(ctx.i18n.t('cmd.my_stats.send_pm'), {
-      reply_to_message_id: ctx.message.message_id,
-    })
-
     const groupAvrg = ctx.groupInfo.stats.textTotal / ctx.groupInfo.stats.messagesCount
     const memberAvrg = ctx.groupMemberInfo.stats.textTotal / ctx.groupMemberInfo.stats.messagesCount
 
     const active = ((ctx.groupMemberInfo.stats.textTotal * 100) / ctx.groupInfo.stats.textTotal).toFixed(2)
     const flood = Math.abs(((memberAvrg - groupAvrg) / groupAvrg) * 100).toFixed(2)
 
-    ctx.telegram.sendMessage(ctx.from.id, ctx.i18n.t('cmd.my_stats.chat', {
+    const pMessage = await ctx.telegram.sendMessage(ctx.from.id, ctx.i18n.t('cmd.my_stats.chat', {
       name: userName(ctx.from, true),
       chatName: ctx.chat.title,
       banTime: humanizeDuration(
@@ -33,11 +29,26 @@ module.exports = async (ctx) => {
       createdAt: dateFormat(ctx.groupMemberInfo.createdAt, 'dd.mm.yyyy H:MM:ss'),
     }), {
       parse_mode: 'HTML',
-    })
+    }).catch(() => {})
 
-    setTimeout(() => {
-      ctx.deleteMessage(message.message_id)
-      ctx.deleteMessage()
-    }, 3 * 1000)
+    let gMessage
+
+    if (pMessage) {
+      gMessage = await ctx.replyWithHTML(ctx.i18n.t('cmd.my_stats.send_pm'), {
+        reply_to_message_id: ctx.message.message_id,
+      })
+    }
+    else {
+      gMessage = await ctx.replyWithHTML(ctx.i18n.t('cmd.my_stats.error.blocked'), {
+        reply_to_message_id: ctx.message.message_id,
+      })
+    }
+
+    if (gMessage) {
+      setTimeout(() => {
+        ctx.deleteMessage(gMessage.message_id)
+        ctx.deleteMessage()
+      }, 3 * 1000)
+    }
   }
 }
